requirements.txt
crewai==0.63.6
crewai-tools==0.8.3
litellm==1.44.22
requests==2.31.0
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
python-dotenv==1.0.0
pydantic==2.5.2
sec-edgar-api==2.1.0
yfinance==0.2.28
beautifulsoup4==4.12.2
lxml==4.9.3
aiohttp==3.9.1
asyncio==3.4.3
python-dateutil==2.8.2
pytz==2023.3
openai==1.3.7
anthropic==0.7.8